[5903] * Fixed: use correct types (mostly in case guid(uint64) and low guid (uint32) correct assigments. Fxied aura loading for pets in case creature casted auras. Fixed item guid creating for loot rolls, this can solve some problem with loot rolling work. Patch provided by MoHax357.

+++ src/game/Pet.cpp
@@ -1207,7 +1207,7 @@ void Pet::_LoadAuras(uint32 timediff)
         do
         {
             Field *fields = result->Fetch();
-            uint32 caster_guid = fields[0].GetUInt64();
+            uint64 caster_guid = fields[0].GetUInt64();
             uint32 spellid = fields[1].GetUInt32();
             uint32 effindex = fields[2].GetUInt32();
             int32 damage     = (int32)fields[3].GetUInt32();
